+++ NEWS.md
@@ -1,7 +1,10 @@
-# CopernicusClimate v0.0.2.0003
+# CopernicusClimate v0.0.2.0004
 
+* New feature:
+  * Translator for Python request: `cds_python_to_r()`
 * Updated documentation
 * Fixes in `cds_build_request()`
+* improved test coverage
 
 # CopernicusClimate v0.0.2